Ceres evaluation bug, but now formatted correctly

+++ src/slam_type_builder.h
@@ -20,11 +20,11 @@ private:
     uint64_t pose_id_ = 0;
     bool odom_initialized_ = false;
     Eigen::Vector2f init_odom_translation_;
-    double init_odom_angle_;
+    float init_odom_angle_;
     Eigen::Vector2f odom_translation_;
-    double odom_angle_;
+    float odom_angle_;
     Eigen::Vector2f last_odom_translation_;
-    double last_odom_angle_;
+    float last_odom_angle_;
     std::vector<slam_types::SLAMNode2D> nodes_;
     std::vector<slam_types::OdometryFactor2D> odom_factors_;
     void AddOdomFactor(std::vector<slam_types::OdometryFactor2D>*);
